refactor(phase-2d-strict): full LoC+file-cap lockdown for Login+PreLogin+AuthDiscovery
Splits the three big mediator phase-action files into co-located
helper modules so every function fits the 10-LoC / 10-statement
ceiling and every file fits the 150-LoC cap.

- Login: LoginPhaseActions.ts now a thin re-export stub.
  New helpers: LoginFormAnchor, LoginSubmitResolve,
  LoginFieldDiscovery, LoginPreOrchestrator, LoginActionExecute,
  LoginFrameScan, LoginUrlHelpers, LoginScopeIntact,
  LoginPostValidate, LoginPostBounce.
- PreLogin: PreLoginPhaseActions slimmed to thin orchestrator +
  re-exports. New helpers: PreLoginRevealDiscovery,
  PreLoginSealedReveal.
- AuthDiscovery: AuthDiscoveryActions slimmed to thin orchestrator.
  New helpers: AuthDiscoveryTelemetry, AuthDiscoveryFinal.
- Small-file extractions in LoginCookieAudit, PostLoginTrafficProbe,
  PreLoginActions, PreLoginRevealProbe, AuthDiscoveryInterstitial.

Strict cluster probe EXIT 0 (37 -> 0 violations). All 22 new/modified
files are inside src/Scrapers/Pipeline/Mediator/{Login,PreLogin,
AuthDiscovery}/.
